requirements.txt
# Core Dependencies
streamlit>=1.28.0
ollama>=0.1.0
faster-whisper>=0.10.0
chromadb>=0.4.15
sqlalchemy>=2.0.23
pydantic>=2.5.0

# Database
aiosqlite>=0.19.0

# PDF Generation
reportlab>=4.0.7
weasyprint>=60.1

# Medical Imaging
pydicom>=2.4.4
pillow>=10.1.0

# NLP & Processing
spacy>=3.7.2
python-dotenv>=1.0.0
pandas>=2.1.3
numpy>=1.26.2

# Audio Processing
pyaudio>=0.2.14
soundfile>=0.12.1

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

